refactor(eth): drop redundant BLOCKHASH slot witnessing and share the witness wire type
Remove the EIP-2935 HistoryStorage slot witnessing from the tx-list
replay. The cross-client stateless executor resolves BLOCKHASH from the
ancestor headers carried in the witness (EIP-2935 does not change the
opcode's semantics), and the BLOCKHASH opcode already records those
headers into the witness during the replay; no Taiko network deploys the
history contract, so the slot reads could never contribute state nodes.
This deletes the OnBlockHashRead tracer hook (and with it the vm.Config
tracer wiring), the block-number bookkeeping, and the post-finalize slot
pull. A new regression test guards what the reference actually consumes:
the resolved ancestor's header must be present in the witness.

Hoist the repeated beacon-root / Prague gates in buildTxListWitness into
locals so the system-caller witness condition visibly mirrors the system
calls it depends on.

Move the debug execution-witness wire type into core/stateless as
stateless.ExecutionWitness, shared by the debug RPC producer and the
fetchpayload consumer, which previously kept a field-for-field duplicate
of the type and its header-decoding converter. debug_executionWitness
error paths now return nil instead of an empty struct alongside the
error.
