+++ eth/taiko_api_debug_test.go
@@ -8,10 +8,7 @@ import (
 	"testing"
 
 	"github.com/ethereum/go-ethereum/core"
-	"github.com/ethereum/go-ethereum/core/stateless"
-	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/core/vm"
-	"github.com/ethereum/go-ethereum/rlp"
 )
 
 func TestDecodeTxListWitnessTxsEmpty(t *testing.T) {
@@ -84,32 +81,5 @@ func TestIsRecoverableNonAnchorTxError(t *testing.T) {
 	}
 	if isRecoverableNonAnchorTxError(fmt.Errorf("wrapped: %w", errors.New("boom"))) {
 		t.Fatalf("wrapped unrelated error must not be recoverable")
-	}
-}
-
-func TestNewExecutionWitness(t *testing.T) {
-	hdr := &types.Header{Number: big.NewInt(7)}
-	w := &stateless.Witness{
-		Headers: []*types.Header{hdr},
-		State:   map[string]struct{}{"node": {}},
-		Codes:   map[string]struct{}{"code": {}},
-		Keys:    map[string]struct{}{"key": {}},
-	}
-	out, err := newExecutionWitness(w)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if len(out.State) != 1 || len(out.Codes) != 1 || len(out.Keys) != 1 || len(out.Headers) != 1 {
-		t.Fatalf("unexpected lengths: %+v", out)
-	}
-	if string(out.State[0]) != "node" || string(out.Codes[0]) != "code" || string(out.Keys[0]) != "key" {
-		t.Fatalf("unexpected witness bytes: %+v", out)
-	}
-	var decoded types.Header
-	if err := rlp.DecodeBytes(out.Headers[0], &decoded); err != nil {
-		t.Fatalf("headers must RLP-decode to a header: %v", err)
-	}
-	if decoded.Number.Uint64() != 7 {
-		t.Fatalf("wrong header decoded: %d", decoded.Number.Uint64())
 	}
 }
